Lazy load PhysicalServer class in spec factory
It seems that FactoryGirl can get into troubles when combined with
Rails' autoloading of classes, which manifests itself with a warning
similar to this: "toplevel constant SomeClassName referenced ...".

To bypass this problem, we force FactoryGirl to lazy load the class by
passing in a string instead of the actual class.