Be super patient and nice when rebooting server
Currently we configure server's PXE boot properties and then reboot it.
Turns out, however, that server reboot operation is not always even
supported. Therefore we now rather stop server and then spin it again.
This way we don't rely on reboot operation support in favor of stop+start.

But beware it's not always trivial to shut server down: we had to handle
four possible power states of the server or else the process would not
succeed.

Signed-off-by: Tadej Borovšak <tadej.borovsak@xlab.si>
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>