Make sure provider_object returns valid object
There is a non-zero chance of ManageIQ database and external Redfish
system to be out-of-sync when the PhysicalServer#provider_object
method is called. In this situation, the current implementation
returns back the nil object, causing hard to debug problems down the
line.

In order to give a sensible error message to the developers, we
changed the provider_object method a bit. This method now raises an
exception of the object cannot be retrieved from the external system.