Update Rails' secrets access in VCR tests
Up until version 5.1.x of Rails, we were able to use symbol or string
to access secret contents. For example, let us assume that we placed
the following content into the `config/secrets.yml` file:

    test:
      redfish:
        some: content

Using Rails 5.0.x, the next two lines return the same value:

    Rails.application.secrets.redfish["some"]
    Rails.application.secrets.redfish[:some]

But when using Rails 5.1.x, the first line is not retrieving the same
value as the second one.